Correct class names to match CSS
The tutorial provides React Component code that lists non-existent CSS classes. This results in improper spacing (see screenshot). It appears the classnames & styles in the CSS file was refactored at some point and this .md guide was not updated. 

The proposed changes use the correct class names here in the markdown.