Update subscribing-to-database-changes.mdx
Include a comment that indicates that this feature is not turned on by default.
Currently a reader may try the code snippet and think that there is a bug in their code when the subscription does not work. In reality they may just need to turn the feature on. This mistake is easy to make as the written documentation does not indicate anywhere that this step must be taken.